FDG-11499 consolidating css
