+++ src/components/data-preview/data-preview-table/data-preview-table.spec.js
@@ -5,7 +5,7 @@ import { RecoilRoot } from 'recoil';
 import { render, within } from '@testing-library/react';
 import DataPreviewTable from './data-preview-table';
 import { DataTableContext } from '../data-preview-context';
-import { contextProps } from '../../table-components/table-test-helper';
+import { contextProps } from '../../table-components/helpers/table-test-helper';
 import { columnsConstructorData } from '../../dtg-table/data-table-helper';
 
 describe('DataPreviewTable component', () => {
