+++ src/components/dtg-table/dtg-table-column-selector.module.scss
@@ -79,6 +79,27 @@
   width: 100%;
 }
 
+.selectColumnsWrapper {
+  display: flex;
+
+  .selectColumnPanelInactive {
+    display: none;
+  }
+
+  .selectColumnPanelActive {
+    display: block;
+    border-left: 1px solid $border-color;
+    margin-top: -2.75rem;
+  }
+
+  .selectColumnPanelActive,
+  .selectColumnPanelInactive {
+    height: 565px;
+    order: 2;
+    width: 450px;
+  }
+}
+
 @media screen and (max-width: $breakpoint-lg - 1) {
   .selectAllContainer {
     .reset {
