+++ src/components/dtg-table/dtg-table.jsx
@@ -3,7 +3,7 @@ import { getDateFilters, getPaginationValues, loadingTimeout, netLoadingDelay } 
 import { defaultPerPageOptions } from '../pagination/pagination-controls';
 import { pagedDatatableRequest, REACT_TABLE_MAX_NON_PAGINATED_SIZE } from '../../utils/api-utils';
 import NotShownMessage from '../dataset-data/table-section-container/not-shown-message/not-shown-message';
-import { loadingIcon, overlay, overlayContainer, overlayContainerNoFooter } from './dtg-table.module.scss';
+import { loadingIcon, overlay, overlayContainer, overlayContainerNoFooter, selectColumnsWrapper } from './dtg-table.module.scss';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { reactTableFilteredDateRangeState } from '../../recoil/reactTableFilteredState';
 import { ErrorBoundary } from 'react-error-boundary';
@@ -11,13 +11,8 @@ import DtgTableApiError from './dtg-table-api-error/dtg-table-api-error';
 import LoadingIndicator from '../loading-indicator/loading-indicator';
 import { getCoreRowModel, getFilteredRowModel, getPaginationRowModel, getSortedRowModel, useReactTable } from '@tanstack/react-table';
 import { columnsConstructorData, constructDefaultColumnsFromTableData, getInvisibleColumns, getSortedColumnsData } from './data-table-helper';
-import {
-  rawDataTableContainer,
-  selectColumnPanelActive,
-  selectColumnPanelInactive,
-  selectColumnsWrapper,
-  tableStyle,
-} from '../table-components/table.module.scss';
+import { selectColumnPanelActive, selectColumnPanelInactive } from './dtg-table-column-selector.module.scss';
+import { rawDataTableContainer, tableStyle } from '../table-components/filtered-table/filtered-table.module.scss';
 
 import TableColumnSelector from '../table-components/column-select/table-column-selector';
 import TableHeader from '../table-components/table-header/table-header';
