+++ src/components/table-components/helpers/table-test-helper.js
@@ -1,66 +1,3 @@
-import { columnsConstructorData } from '../dtg-table/data-table-helper';
-
-export const mockMeta = {
-  count: 2,
-  labels: {
-    record_date: 'Record Date',
-    debt_held_public_amt: 'Debt Held by the Public',
-    intragov_hold_amt: 'Intragovernmental Holdings',
-    tot_pub_debt_out_amt: 'Total Public Debt Outstanding',
-    src_line_nbr: 'Source Line Number',
-    record_fiscal_year: 'Fiscal Year',
-    record_fiscal_quarter: 'Fiscal Quarter Number',
-    record_calendar_year: 'Calendar Year',
-    record_calendar_quarter: 'Calendar Quarter Number',
-    record_calendar_month: 'Calendar Month Number',
-    record_calendar_day: 'Calendar Day Number',
-    mock_small_fraction: 'Mock Small Fraction',
-    mock_percent_string: 'Mock Percent String',
-    mock_percent: 'Mock Percent',
-    negative_currency: 'Negative Currency',
-    daily_index: 'Daily Index',
-    daily_int_accrual_rate: 'Daily Accrual Rate',
-    spread: 'Spread',
-    additional_date: 'Additional Date',
-  },
-  dataTypes: {
-    record_date: 'DATE',
-    debt_held_public_amt: 'CURRENCY',
-    intragov_hold_amt: 'CURRENCY3',
-    tot_pub_debt_out_amt: 'CURRENCY',
-    src_line_nbr: 'NUMBER',
-    record_fiscal_year: 'YEAR',
-    record_fiscal_quarter: 'QUARTER',
-    record_calendar_year: 'YEAR',
-    record_calendar_quarter: 'QUARTER',
-    record_calendar_month: 'MONTH',
-    record_calendar_day: 'DAY',
-    mock_small_fraction: 'SMALL_FRACTION',
-    mock_percent_string: 'STRING',
-    mock_percent: 'PERCENTAGE',
-    negative_currency: 'CURRENCY3',
-    daily_index: 'NUMBER',
-    daily_int_accrual_rate: 'NUMBER',
-    spread: 'NUMBER',
-    additional_date: 'STRING',
-  },
-  dataFormats: {
-    record_date: 'YYYY-MM-DD',
-    debt_held_public_amt: '$10.20',
-    intragov_hold_amt: '$10.20',
-    tot_pub_debt_out_amt: '$10.20',
-    src_line_nbr: '10.2',
-    record_fiscal_year: 'YYYY',
-    record_fiscal_quarter: 'Q',
-    record_calendar_year: 'YYYY',
-    record_calendar_quarter: 'Q',
-    record_calendar_month: 'MM',
-    record_calendar_day: 'DD',
-  },
-  'total-count': 2,
-  'total-pages': 1,
-};
-
 export const mockColumnConfig = [
   {
     property: 'record_date',
