--- src/components/table-components/table.module.scss
@@ -1,120 +1,0 @@
-@use 'src/variables.module' as *;
-
-$borderRadius: 5px;
-
-.tableStyle {
-  overflow-x: auto;
-  width: 100%;
-  order: 1;
-
-  .rawDataTableContainer,
-  .nonRawDataTableContainer {
-    position: relative;
-    box-sizing: border-box;
-    border-top: 0;
-    background-color: white;
-    overflow-y: auto;
-  }
-
-  .rawDataTableContainer {
-    height: 521px;
-  }
-  .nonRawDataTableContainer {
-    max-height: 521px;
-  }
-
-  table {
-    margin-top: -3px;
-  }
-
-  table,
-  .divTable {
-    border-bottom: 1px solid lightgray;
-    width: fit-content;
-  }
-
-  th,
-  .th {
-    cursor: default;
-    background-color: white;
-    position: relative;
-    font-weight: bold;
-    text-align: center;
-    height: 1.875rem;
-    padding: 0.25rem 0.5rem 0.75rem;
-  }
-
-  .tr {
-    display: flex;
-  }
-
-  tr,
-  .tr {
-    width: fit-content;
-    height: 1.875rem;
-  }
-
-  td,
-  .td {
-    padding: 0.25rem 0.5rem;
-  }
-
-  #th-parent {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-  }
-
-  td,
-  .td {
-    height: 1.875rem;
-  }
-}
-.range {
-  font-weight: $semi-bold-weight;
-}
-
-.overlayContainer,
-.overlayContainerNoFooter {
-  position: relative;
-}
-
-.overlayContainerNoFooter {
-  border: 1px solid $dd-border-color;
-  border-top: 0;
-  border-bottom-left-radius: $borderRadius;
-  border-bottom-right-radius: $borderRadius;
-}
-.overlayContainerNoFooterChart {
-  border: none;
-}
-.selectColumnsWrapper {
-  display: flex;
-
-  .selectColumnPanelInactive {
-    display: none;
-  }
-
-  .selectColumnPanelActive {
-    display: block;
-    border-left: 1px solid $border-color;
-    margin-top: -2.75rem;
-  }
-
-  .selectColumnPanelActive,
-  .selectColumnPanelInactive {
-    height: 565px;
-    order: 2;
-    width: 450px;
-  }
-}
-
-.downloadLinkContainer {
-  display: flex;
-  flex-direction: row;
-  gap: 0.5rem;
-}
-
-.downloadLinkIcon {
-  color: $primary;
-}
